refactor(fork_engine): completely reorganize code
